+++ src/service/skillService/skill.service.ts
@@ -1,12 +1,19 @@
-import { ObjectId } from "mongoose";
+import mongoose, { ObjectId } from "mongoose";
 import { SkillInput, skillModal } from "../../schema/skillsSchema/skill.schema";
+import { employeeSkillModal } from "../../schema/employeeSkill/employeeSkill.schema";
 
 
 export class SkillService{
     async newSkill(input: SkillInput, _id){
         try{
-            const skill = await skillModal.create({...input, createdBy:_id});
-            return skill
+            const isExist  = await skillModal.findOne({$and:[{skill:input?.skill}, {categoryId: new mongoose.Types.ObjectId(input?.categoryId)}]});
+            if(isExist){
+                throw new Error("Skill Already Exist");   
+            }else{
+
+                const skill = await skillModal.create({...input, createdBy:_id});
+                return skill
+            }
             
         }catch(err){
             throw new Error(err.message);
@@ -47,7 +54,7 @@ export class SkillService{
 
             ]);
 
-            return skills?.map((sc)=>({...sc, categoryId: sc?.category[0]?.category, createdBy: sc?.createdByDetail[0]?.firstname+" "+sc?.createdByDetail[0]?.lastname}));
+            return skills?.map((sc)=>({...sc, categoryName: sc?.category[0]?.category, createdBy: sc?.createdByDetail[0]?.firstname+" "+sc?.createdByDetail[0]?.lastname}));
         }catch(err){
             throw new Error(err.message);
         }
@@ -64,9 +71,17 @@ export class SkillService{
 
     async deleteSkill(_ids){
         try{
-            await skillModal.deleteMany({_id:{$in:_ids}});
-            return {
-                message: "Skills deleted successfully"
+            // check if this skill is in user
+            const isInUser = await employeeSkillModal?.find({skill:{$in:_ids}});
+            console.log(isInUser, "isInUser");
+            if(isInUser?.length>0){
+                throw new Error("Action omitted ! This skill is in use");
+            }else{
+
+                await skillModal.deleteMany({_id:{$in:_ids}});
+                return {
+                    message: "Skills deleted successfully"
+                }
             }
         }
         catch(err){
